+++ net.resheim.eclipse.timekeeper.test/src/main/java/net/resheim/eclipse/timekeeper/ui/test/TestUtility.java
@@ -70,7 +70,7 @@ public class TestUtility {
 		});
 		ITask task = new LocalTask(id, text);
 		tl.addTask(task, c);
-		return new Task(task);
+		return TimekeeperPlugin.getDefault().getTask(task);
 	}
 
 	/**